# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Aspire API",
        "version": "0.2"
    },
    "paths": {
        "/api/loans": {
            "get": {
                "tags": [
                    "Loan"
                ],
                "summary": "Gets all the authenticaed users loan",
                "description": "Returns Json of all loan records by the user",
                "operationId": "index",
                "responses": {
                    "200": {
                        "description": "Record(s) found",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Loan"
                ],
                "summary": "Create a loan record",
                "description": "Post data / loan application by the authenticated user",
                "operationId": "store",
                "parameters": [
                    {
                        "name": "description",
                        "in": "query",
                        "description": "description",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "amount",
                        "in": "query",
                        "description": "amount to borrow",
                        "required": true,
                        "schema": {
                            "type": "number"
                        }
                    },
                    {
                        "name": "terms",
                        "in": "query",
                        "description": "in weeks (fixed interest rate to 5% per week)",
                        "required": true,
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Record Successfully saved."
                    },
                    "422": {
                        "description": "Unprocessable Entity"
                    },
                    "500": {
                        "description": "Code Error"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            }
        },
        "/api/loans/{loan}": {
            "patch": {
                "tags": [
                    "Loan"
                ],
                "summary": "Update a loan record",
                "description": "Patch data / loan application by the authenticated user",
                "operationId": "update",
                "parameters": [
                    {
                        "name": "loan",
                        "in": "path",
                        "description": "Loan Id",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "description",
                        "in": "query",
                        "description": "description",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "amount",
                        "in": "query",
                        "description": "amount to borrow",
                        "required": true,
                        "schema": {
                            "type": "number"
                        }
                    },
                    {
                        "name": "terms",
                        "in": "query",
                        "description": "in weeks (fixed interest rate to 5% per week)",
                        "required": true,
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Record Successfully updated."
                    },
                    "422": {
                        "description": "Unprocessable Entity"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "500": {
                        "description": "Code Error"
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            }
        },
        "/api/loans/{loan}/approve": {
            "patch": {
                "tags": [
                    "Loan"
                ],
                "summary": "Approve a loan record",
                "description": "Approve the pending loan application by the authenticated user",
                "operationId": "approve",
                "parameters": [
                    {
                        "name": "loan",
                        "in": "path",
                        "description": "Loan Id",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Record Successfully updated."
                    },
                    "500": {
                        "description": "Code Error"
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            }
        },
        "/api/loans/{loan}/full/{terms?}": {
            "post": {
                "tags": [
                    "Loan"
                ],
                "summary": "pays the loan in full base on terms",
                "description": "Pay the loan in full base on terms (default value of terms is 1)",
                "operationId": "fullRepayment",
                "parameters": [
                    {
                        "name": "loan",
                        "in": "path",
                        "description": "Loan Id",
                        "required": true,
                        "schema": {
                            "type": "number"
                        }
                    },
                    {
                        "name": "terms",
                        "in": "path",
                        "description": "in weeks (fixed interest rate to 5% per week)",
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Fully Paid."
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "500": {
                        "description": "Code Error"
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            }
        },
        "/api/loans/{loan}/pay": {
            "post": {
                "tags": [
                    "Loan"
                ],
                "summary": "pays the loan base on terms",
                "description": "Pay the loan base on terms (assumed that it is a weekly repayment)",
                "operationId": "repayment",
                "parameters": [
                    {
                        "name": "loan",
                        "in": "path",
                        "description": "Loan Id",
                        "required": true,
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Paid."
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "500": {
                        "description": "Code Error"
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "passport": {
                "type": "oauth2",
                "description": "Laravel passport oauth2 security.",
                "in": "header",
                "scheme": "https",
                "flows": {
                    "password": {
                        "authorizationUrl": "http://127.0.0.1:8000/oauth/authorize",
                        "tokenUrl": "http://127.0.0.1:8000/oauth/token",
                        "refreshUrl": "http://127.0.0.1:8000/token/refresh",
                        "scopes": []
                    }
                }
            }
        }
    },
    "security": [
        {
            "passport": []
        }
    ]
}